+++ sharding-jdbc-core/src/main/java/com/dangdang/ddframe/rdb/sharding/executor/wrapper/AbstractExecutorWrapper.java
@@ -18,6 +18,7 @@
 package com.dangdang.ddframe.rdb.sharding.executor.wrapper;
 
 import com.dangdang.ddframe.rdb.sharding.executor.event.DMLExecutionEvent;
+import com.dangdang.ddframe.rdb.sharding.executor.event.DQLExecutionEvent;
 import com.dangdang.ddframe.rdb.sharding.router.SQLExecutionUnit;
 import com.google.common.base.Optional;
 import lombok.Getter;
@@ -39,9 +40,18 @@ public abstract class AbstractExecutorWrapper {
      * 
      * @return 是否为DML语句
      */
-    public final boolean isDML() {
+    final boolean isDML() {
         String sql = sqlExecutionUnit.getSql();
         return sql.toLowerCase().startsWith("insert") || sql.toLowerCase().startsWith("update") || sql.toLowerCase().startsWith("delete");
+    }
+    
+    /**
+     * 判断SQL是否为DQL语句.
+     * 
+     * @return 是否为DQL语句
+     */
+    final boolean isDQL() {
+        return sqlExecutionUnit.getSql().toLowerCase().startsWith("select");
     }
     
     /**
@@ -50,4 +60,11 @@ public abstract class AbstractExecutorWrapper {
      * @return DML类SQL执行时事件
      */
     public abstract Optional<DMLExecutionEvent> getDMLExecutionEvent();
+    
+    /**
+     * 获取DML类SQL执行时事件.
+     * 
+     * @return DQL类SQL执行时事件
+     */
+    public abstract Optional<DQLExecutionEvent> getDQLExecutionEvent();
 }