+++ sharding-jdbc-core/src/test/java/com/dangdang/ddframe/rdb/integrate/masterslave/pstatement/ShardingMasterSlaveForPStatementWithDMLTest.java
@@ -18,6 +18,7 @@
 package com.dangdang.ddframe.rdb.integrate.masterslave.pstatement;
 
 import com.dangdang.ddframe.rdb.integrate.masterslave.AbstractShardingMasterSlaveDBUnitTest;
+import com.dangdang.ddframe.rdb.sharding.api.HintManager;
 import com.dangdang.ddframe.rdb.sharding.jdbc.ShardingDataSource;
 import com.dangdang.ddframe.rdb.sharding.parser.result.router.SQLStatementType;
 import org.dbunit.DatabaseUnitException;
@@ -55,6 +56,31 @@ public final class ShardingMasterSlaveForPStatementWithDMLTest extends AbstractS
             }
         }
         assertDataSet("insert", "insert");
+    }
+    
+    @Test
+    public void assertInsertWithHint() throws SQLException, DatabaseUnitException {
+        String sql = "INSERT INTO `t_order` (`order_id`, `user_id`, `status`) VALUES (?, ?, ?)";
+        for (int i = 1; i <= 10; i++) {
+            for (int j = 1; j <= 10; j++) {
+                try (
+                        Connection connection = shardingDataSource.getConnection();
+                        HintManager hintManager = HintManager.getInstance()
+                ) {
+                    hintManager.addDatabaseShardingValue("t_order", "user_id", j);
+                    hintManager.addTableShardingValue("t_order", "order_id", i);
+                    PreparedStatement preparedStatement = connection.prepareStatement(sql);
+                    preparedStatement.setInt(1, i);
+                    preparedStatement.setInt(2, j);
+                    preparedStatement.setString(3, "insert");
+                    preparedStatement.executeUpdate();
+                }
+            }
+        }
+        try (HintManager hintManager = HintManager.getInstance()) {
+            hintManager.setMasterRouteOnly();
+            assertDataSet("insert", "insert");
+        }
     }
     
     @Test